requirements.txt
astroid==2.5.1
attrs==20.3.0
autopep8==1.5.6
click==7.1.2
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
iniconfig==1.1.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
motor==2.3.1
packaging==20.9
passlib==1.7.4
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pydantic==1.8.1
pylint==2.7.2
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.2.2
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.7.2
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
werkzeug==1.0.1
wrapt==1.12.1
